+++ dev/doc/changes.md
@@ -1,3 +1,10 @@
+## Changes between Coq 8.14 and Coq 8.15
+
+### Universes
+
+- Type `Univ.UContext` now embeds universe user names, generally
+  resulting in more concise code.
+
 ## Changes between Coq 8.13 and Coq 8.14
 
 ### Build system and library infrastructure